feat: rename libs ::by sergio giraldo @ 20231204T1318CET, gpg signed
