+++ solutions/day4/solution.js
@@ -8,7 +8,7 @@ const report = (...messages) => {
     streamAnswer.write(JSON.stringify(messages) + "\n");
     console.log(`[${require(fromHere("../../package.json")).logName} / ${__dirname.split(path.sep).pop()}]`, ...messages);
   };
-const lib = require("../lib/004");
+const lib = require("../lib/day4");
 
  function run () {
     solveForFirstStar();
