+++ solutions/lib/day2.js
