+++ solutions/lib/day3.js
