+++ solutions/lib/day4.js
